+++ NEWS.md
@@ -16,6 +16,7 @@ NEW FEATURES :
 * `updateScenarioBuilder()` New type of series "hfl" ("hydro final level", similar to "hydrolevels") is available
 * `editBindingConstraint()` : control the dimensions of the matrix only if a time series is provided by the user for optimization
 * `.createCluster()` uses a specific endpoint to write cluster's metadata and commands to write matrix
+* Add new function `setThematicTrimming()` to set the thematic trimming in file `generaldata.ini`
 
 
 ### Breaking changes  :